+++ soupsieve/css_parser.py
@@ -23,17 +23,10 @@ PAT_ID = r'#(?:[-\w]|{esc})+'.format(esc=CSS_ESCAPES)
 
 PAT_CLASS = r'\.(?:[-\w]|{esc})+'.format(esc=CSS_ESCAPES)
 
-PAT_HTML_TAG = r'(?:(?:(?:[-\w]|{esc})+|\*)?\|)?(?:(?:[-\w]|{esc})+|\*)'.format(esc=CSS_ESCAPES)
+PAT_TAG = r'(?:(?:(?:[-\w]|{esc})+|\*)?\|)?(?:(?:[-\w]|{esc})+|\*)'.format(esc=CSS_ESCAPES)
 
-PAT_XML_TAG = r'(?:(?:(?:[-\w.]|{esc})+|\*)?\|)?(?:(?:[-\w.]|{esc})+|\*)'.format(esc=CSS_ESCAPES)
-
-PAT_HTML_ATTR = r'''(?x)
+PAT_ATTR = r'''(?x)
 \[{ws}*(?P<ns_attr>(?:(?:(?:[-\w]|{esc})+|\*)?\|)?(?:[-\w]|{esc})+)
-{attr}
-'''.format(ws=WS, esc=CSS_ESCAPES, attr=ATTR)
-
-PAT_XML_ATTR = r'''(?x)
-\[{ws}*(?P<ns_attr>(?:(?:(?:[-\w]|{esc})+|\*)?\|)?(?:[-\w.]|{esc})+)
 {attr}
 '''.format(ws=WS, esc=CSS_ESCAPES, attr=ATTR)
 
@@ -113,24 +106,6 @@ class SelectorPattern:
         return True
 
 
-class HtmlSelectorPattern(SelectorPattern):
-    """HTML selector pattern."""
-
-    def enabled(self, flags):
-        """Enabled."""
-
-        return (flags & util.MODE_MSK) in (util.HTML, util.HTML5, util.XHTML)
-
-
-class XmlSelectorPattern(SelectorPattern):
-    """XML selector pattern."""
-
-    def enabled(self, flags):
-        """Enabled."""
-
-        return (flags & util.MODE_MSK) in (util.XML,)
-
-
 class _Selector:
     """
     Intermediate selector class.
@@ -206,12 +181,10 @@ class CSSParser:
             ("contains", SelectorPattern(PAT_CONTAINS)),
             ("pseudo_nth_child", SelectorPattern(PAT_PSEUDO_NTH_CHILD)),
             ("pseudo_nth_type", SelectorPattern(PAT_PSEUDO_NTH_TYPE)),
-            ("id", HtmlSelectorPattern(PAT_ID)),
-            ("class", HtmlSelectorPattern(PAT_CLASS)),
-            ("html_tag", HtmlSelectorPattern(PAT_HTML_TAG)),
-            ("xml_tag", XmlSelectorPattern(PAT_XML_TAG)),
-            ("html_attribute", HtmlSelectorPattern(PAT_HTML_ATTR)),
-            ("xml_attribute", XmlSelectorPattern(PAT_XML_ATTR)),
+            ("id", SelectorPattern(PAT_ID)),
+            ("class", SelectorPattern(PAT_CLASS)),
+            ("tag", SelectorPattern(PAT_TAG)),
+            ("attribute", SelectorPattern(PAT_ATTR)),
             ("pseudo_close", SelectorPattern(PAT_PSEUDO_CLOSE)),
             ("combine", SelectorPattern(PAT_SPLIT))
         ]
@@ -222,13 +195,11 @@ class CSSParser:
 
         self.pattern = selector
         self.flags = flags
-        mode = flags & util.MODE_MSK
-
-        if mode in (util.HTML, util.HTML5, util.XML, util.XHTML, 0):
-            self.mode = mode if mode else util.DEFAULT_MODE
-        else:
-            raise ValueError("Invalid SelectorMatcher flag(s) '{}'".format(mode))
-        self.adjusted_flags = flags | self.mode
+        dflags = self.flags & util.DEPRECATED_FLAGS
+        if dflags:
+            util.warn_deprecated(
+                "The following flags are deprecated and may be repurposed in the future '0x%02X'" % dflags
+            )
 
     def parse_attribute_selector(self, sel, m, has_selector):
         """Create attribute selector from the returned regex match."""
@@ -236,6 +207,8 @@ class CSSParser:
         case = util.lower(m.group('case').strip()) if m.group('case') else None
         parts = [css_unescape(a.strip()) for a in m.group('ns_attr').split('|')]
         ns = ''
+        is_type = False
+        pattern2 = None
         if len(parts) > 1:
             ns = parts[0]
             attr = parts[1]
@@ -243,10 +216,12 @@ class CSSParser:
             attr = parts[0]
         if case:
             flags = re.I if case == 'i' else 0
-        elif self.mode == util.XML:
-            flags = 0
+        elif util.lower(attr) == 'type':
+            flags = re.I
+            is_type = True
         else:
-            flags = re.I if util.lower(attr) == 'type' and not ns else 0
+            flags = 0
+
         op = m.group('cmp')
         if op:
             value = css_unescape(
@@ -275,8 +250,10 @@ class CSSParser:
         else:
             # Value matches
             pattern = re.compile(r'^%s$' % re.escape(value), flags)
+        if is_type:
+            pattern2 = re.compile(pattern.pattern)
         has_selector = True
-        sel.attributes.append(ct.SelectorAttribute(attr, ns, pattern))
+        sel.attributes.append(ct.SelectorAttribute(attr, ns, pattern, pattern2))
         return has_selector
 
     def parse_tag_pattern(self, sel, m, has_selector):
@@ -513,9 +490,9 @@ class CSSParser:
                         has_selector, sel = self.parse_split(sel, m, has_selector, selectors, relations, is_pseudo)
                     split_last = True
                     continue
-                elif key in ('html_attribute', 'xml_attribute'):
+                elif key == 'attribute':
                     has_selector = self.parse_attribute_selector(sel, m, has_selector)
-                elif key in ('html_tag', 'xml_tag'):
+                elif key == 'tag':
                     if has_selector:
                         raise SyntaxError("Tag must come first")
                     has_selector = self.parse_tag_pattern(sel, m, has_selector)
@@ -556,7 +533,7 @@ class CSSParser:
         while index <= end:
             m = None
             for k, v in self.css_tokens.items():
-                if not v.enabled(self.adjusted_flags):
+                if not v.enabled(self.flags):  # pragma: no cover
                     continue
                 m = v.pattern.match(pattern, index)
                 if m: