+++ tests/test_level2.py
@@ -23,7 +23,6 @@ Not supported (with current opinions or plans the matter):
 - `:focus`: Items cannot be focused in our environment, so this has little meaning and will not be implemented.
 """
 from . import util
-import soupsieve as sv
 
 
 class TestLevel2(util.TestCase):
@@ -50,7 +49,7 @@ class TestLevel2(util.TestCase):
             markup,
             "div > span",
             ["3"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # No spaces
@@ -58,7 +57,7 @@ class TestLevel2(util.TestCase):
             markup,
             "div>span",
             ["3"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_direct_sibling(self):
@@ -82,7 +81,7 @@ class TestLevel2(util.TestCase):
             markup,
             "span + span",
             ["5", "6"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # No spaces
@@ -90,7 +89,7 @@ class TestLevel2(util.TestCase):
             markup,
             "span+span",
             ["5", "6"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Complex
@@ -98,7 +97,7 @@ class TestLevel2(util.TestCase):
             markup,
             "span#4 + span#5",
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_wild_tag(self):
@@ -119,7 +118,7 @@ class TestLevel2(util.TestCase):
             """,
             "body *",
             ["0", "1", "2", "3", "4", "5", "6", "div", "pre"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_attribute(self):
@@ -142,7 +141,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[href]",
             ["2"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # With spaces
@@ -150,7 +149,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[   href   ]",
             ["2"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_multi_attribute(self):
@@ -172,7 +171,7 @@ class TestLevel2(util.TestCase):
             """,
             "span[id].test[data-test=test]",
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_attribute_equal(self):
@@ -196,7 +195,7 @@ class TestLevel2(util.TestCase):
             markup,
             '[id=5]',
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Single quoted
@@ -204,7 +203,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[id='5']",
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Double quoted
@@ -212,7 +211,7 @@ class TestLevel2(util.TestCase):
             markup,
             '[id="5"]',
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # With spaces
@@ -220,35 +219,35 @@ class TestLevel2(util.TestCase):
             markup,
             '[  id  =  "5"  ]',
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             '[ID="5"]',
             ["5"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             '[  id  =  "5"  ]',
             ["5"],
-            flags=sv.HTML
+            flags=util.HTML
         )
 
         self.assert_selector(
             markup,
             '[ID="5"]',
             ["5"],
-            flags=sv.HTML
+            flags=util.HTML
         )
 
         self.assert_selector(
             '<span bad="5"></span>',
             '[  id  =  "5"  ]',
             [],
-            flags=sv.HTML
+            flags=util.HTML
         )
 
     def test_attribute_type(self):
@@ -275,14 +274,14 @@ class TestLevel2(util.TestCase):
             markup,
             '[type="test"]',
             ["0", '2'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             '[type="test"]',
             ['2'],
-            flags=sv.XML
+            flags=util.XML
         )
 
     def test_attribute_start_dash(self):
@@ -303,7 +302,7 @@ class TestLevel2(util.TestCase):
             """,
             "[lang|=en]",
             ["0"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_attribute_contains_space(self):
@@ -327,7 +326,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[class~=test2]",
             ["0"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Start of list
@@ -335,7 +334,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[class~=test-a]",
             ["pre"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # End of list
@@ -343,7 +342,7 @@ class TestLevel2(util.TestCase):
             markup,
             "[class~=test-b]",
             ["pre"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_first_child(self):
@@ -364,5 +363,5 @@ class TestLevel2(util.TestCase):
             """,
             "span:first-child",
             ["1", "4"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )