+++ tests/test_level4.py
@@ -95,14 +95,14 @@ class TestLevel4(util.TestCase):
             markup,
             "[class*=WORDS]",
             [],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             "[class*=WORDS i]",
             ["0", "3", "pre"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         with self.assertRaises(SyntaxError):
@@ -132,14 +132,14 @@ class TestLevel4(util.TestCase):
             markup,
             '[type="test" s]',
             ['2'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             '[type="test" i]',
             ['0', '2'],
-            flags=sv.XML
+            flags=util.XML
         )
 
     def test_is_matches_where(self):
@@ -157,21 +157,21 @@ class TestLevel4(util.TestCase):
             markup,
             ":is(span, a)",
             ["1", "2"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             ":is(span, a:matches(#2))",
             ["1", "2"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             ":where(span, a:matches(#2))",
             ["1", "2"],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Each pseudo class is evaluated separately
@@ -180,7 +180,7 @@ class TestLevel4(util.TestCase):
             markup,
             ":is(span):not(span)",
             [],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Each pseudo class is evaluated separately
@@ -189,7 +189,7 @@ class TestLevel4(util.TestCase):
             markup,
             ":is(span):is(div)",
             [],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         # Each pseudo class is evaluated separately
@@ -198,7 +198,7 @@ class TestLevel4(util.TestCase):
             markup,
             ":is(a):is(#2)",
             ['2'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_multi_nested_not(self):
@@ -221,7 +221,7 @@ class TestLevel4(util.TestCase):
             markup,
             'div :not(p, :not([id=5]))',
             ['5'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_has(self):
@@ -270,49 +270,49 @@ class TestLevel4(util.TestCase):
             markup,
             'div:not(.aaaa):has(.kkkk > p.llll)',
             ['4', '5', '6'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             'div:NOT(.aaaa):HAS(.kkkk > p.llll)',
             ['4', '5', '6'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             'p:has(+ .dddd:has(+ div .jjjj))',
             ['2'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             'p:has(~ .jjjj)',
             ['7', '8'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup2,
             'div:has(> .bbbb, .ffff, .jjjj)',
             ['0', '4', '8'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup2,
             'div:has(> :not(.bbbb, .ffff, .jjjj))',
             ['2', '6', '8'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup2,
             'div:not(:has(> .bbbb, .ffff, .jjjj))',
             ['2', '6'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
     def test_nth_child_of_s(self):
@@ -337,12 +337,12 @@ class TestLevel4(util.TestCase):
             markup,
             ":nth-child(2n + 1 of :is(p, span).test)",
             ['2', '6', '10'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )
 
         self.assert_selector(
             markup,
             ":nth-child(-n+3 of p)",
             ['0', '1', '7'],
-            flags=sv.HTML5
+            flags=util.HTML5
         )