+++ tests/util.py
@@ -4,6 +4,11 @@ import bs4
 import textwrap
 import soupsieve as sv
 
+HTML5 = 1
+HTML = 2
+XHTML = 4
+XML = 8
+
 
 class TestCase(unittest.TestCase):
     """Test case."""
@@ -11,12 +16,12 @@ class TestCase(unittest.TestCase):
     def assert_selector(self, markup, selectors, expected_ids, namespaces={}, flags=0):
         """Assert selector."""
 
-        mode = flags & sv.util.MODE_MSK
-        if mode == sv.HTML:
+        mode = flags & 0x0F
+        if mode == HTML:
             bs_mode = 'lxml'
-        elif mode in (sv.HTML5, 0):
+        elif mode in (HTML5, 0):
             bs_mode = 'html5lib'
-        elif mode in (sv.XHTML, sv.XML):
+        elif mode in (XHTML, XML):
             bs_mode = 'xml'
         soup = bs4.BeautifulSoup(textwrap.dedent(markup.replace('\r\n', '\n')), bs_mode)